// app/src/components/Sidebar/types.ts
export type SortBy = 'name' | 'createdAt';
export type JoinStatus = 'accepting-members' | 'not-accepting-members';
export type MembershipProcess =
  | 'open-membership'
  | 'audition-required'
  | 'application-required';
export type MemberCount =
  | 'less-than-20'
  | '20-50'
  | '50-100'
  | '100-150'
  | 'more-than-150';